Moved NodeKeys to a separate module.
