+++ optihood/entities.py
@@ -1,0 +1,13 @@
+import enum as _enum
+
+
+class NodeKeys(_enum.StrEnum):
+    buses = "buses"
+    grid_connection = "grid_connection"
+    commodity_sources = "commodity_sources"
+    solar = "solar"
+    transformers = "transformers"
+    demand = "demand"
+    storages = "storages"
+    stratified_storage = "stratified_storage"
+    profiles = "profiles"
